applied version controls to footer sections
Applied `shouldDisplay` property to each section header, to be as relaxed as the items it holds (ie. if it contains an item with `shouldDisplay: true` then the category will also be `true`, if all items restricted to `contentVersion > 1` then section will use this. Applied conditional in return statement to only render section if shouldDisplay === true.