Remove Trinity, add more info
Trinity is not production ready client and is not being developed anymore. Having it listed with other clients here confuses users and is inconsistent with statement above the table: 'All of them are actively worked on and pass client tests.'

Also added more info about other clients to balance it and fixed some typos.