Add description/purpose to Eth R&D Discord link
Quite a few people seem to come to the Eth R&D Discord asking support
questions, even though that is explicitly not allowed in that server
according to its #welcome message.

One possible reason this happens is that it is the first Discord/chat
entry they see on the "community" page. To redirect users to a more
appropriate chat room/server we:

- Elaborate that the Eth R&D Discord is _not_ for support questions, and
  only for focused technical discussions.

- move it down on the list, such that it is not the first link people
  see if they are looking for a chat room to ask for support in.