Change author
Hey, following up on [this](https://github.com/ethereum/ethereum-org-website/pull/2653). My job while interning at Alchemy was just to put up the existing NFT Minter tutorial from our website on the Ethereum Foundation. After looking at the thread of comments a few months later on the linked PR, @smudgil is absolutely right — she is the rightful owner of this tutorial, and I had no clue!

I've gone ahead and taken down the Twitter post that Sumi linked, and in addition, this is a PR with a request to change the top level name to Sumi's GitHub. Please see that this PR is merged, and Sumi becomes the official rightful owner here.

Thank you!!