refactor(cache): remove unnecessary failover
readFromCache returns the content of the local storage if it exists.
the failover is already in the parsing methods as it contains at least
the item of the local storage or a default one as a fallback.
Anyway, the JSON.parse returns a CacheStorage<T>. We do not want the
compiler to interpret as a any type eventhough the return is a
CacheStorage<T>.

Refs:#6392