update page-layer-2.json
So I checked a little and I found out that the link provided was wrong.
The "path" given to the Sharding topic page in **href = "path"** was not correct. 
I've changed it to the correct one. It will work now. 

The correct link:     https://ethereum.org/en/upgrades/sharding/   
The broken link:     https://ethereum.org/upgrades/sharding/

Try clicking on them.